Added support for other Chromium browsers than Google Chrome

+++ NativeMessaging/Host.cs
@@ -1,38 +1,34 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Reflection;
 using System.IO;
-using Microsoft.Win32;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using System.Threading;
 
-namespace NativeMessaging
-{
+namespace NativeMessaging {
     /// <summary>
     /// Abstract class that should be extended to communicate with Chrome
     /// </summary>
-    public abstract class Host
-    {
+    public abstract class Host {
+        private readonly bool SendConfirmationReceipt;
+        private readonly string ManifestPath;
+
         /// <summary>
         /// Name of the Native Messaging Host
         /// </summary>
         public abstract string Hostname { get; }
 
-        private readonly bool SendConfirmationReceipt;
-        private readonly string ManifestPath;
-
-        private const string RegKeyBaseLocation = "SOFTWARE\\Google\\Chrome\\NativeMessagingHosts\\";
+        /// <summary>
+        /// List of supported chromium Browsers.
+        /// </summary>
+        public List<ChromiumBrowser> SupportedBrowsers { get; }
 
         /// <summary>
         /// Creates the Host Object
         /// </summary>
         /// <param name="sendConfirmationReceipt"><see langword="true" /> for the host to automatically send message confirmation receipt.</param>
-        public Host(bool sendConfirmationReceipt = true)
-        {
+        public Host(bool sendConfirmationReceipt = true) {
+            SupportedBrowsers = new List<ChromiumBrowser>(2);
+
             SendConfirmationReceipt = sendConfirmationReceipt;
             ManifestPath = Path.Combine(Utils.AssemblyLoadDirectory(), Hostname + "-manifest.json");
         }
@@ -40,21 +36,23 @@ namespace NativeMessaging
         /// <summary>
         /// Starts listening for input.
         /// </summary>
-        public void Listen()
-        {
-            if (!IsRegisteredWithChrome()) throw new NotRegisteredWithChromeException(Hostname);
+        public void Listen() {
+            if (!IsRegistered())
+                throw new NotRegisteredWithBrowserException(Hostname);
+
             JObject data;
-            while ((data = Read()) != null)
-            {
-                Utils.LogMessage("Data Received:" + JsonConvert.SerializeObject(data));
-                if (SendConfirmationReceipt) SendMessage(new ResponseConfirmation(data).GetJObject());
+            while ((data = Read()) != null) {
+                Log.LogMessage("Data Received:" + JsonConvert.SerializeObject(data));
+
+                if (SendConfirmationReceipt)
+                    SendMessage(new ResponseConfirmation(data).GetJObject());
                 ProcessReceivedMessage(data);
             }
         }
 
-        private JObject Read()
-        {
-            Utils.LogMessage("Waiting for Data");
+        private JObject Read() {
+            Log.LogMessage("Waiting for Data");
+
             var stdin = Console.OpenStandardInput();
 
             var lengthBytes = new byte[4];
@@ -62,7 +60,9 @@ namespace NativeMessaging
 
             var buffer = new char[BitConverter.ToInt32(lengthBytes, 0)];
 
-            using (var reader = new StreamReader(stdin)) while (reader.Peek() >= 0) reader.Read(buffer, 0, buffer.Length);
+            using (var reader = new StreamReader(stdin))
+                while (reader.Peek() >= 0)
+                    reader.Read(buffer, 0, buffer.Length);
 
             return JsonConvert.DeserializeObject<JObject>(new string(buffer));
         }
@@ -71,9 +71,9 @@ namespace NativeMessaging
         /// Sends a message to Chrome, note that the message might not be able to reach Chrome if the stdIn / stdOut aren't properly configured (i.e. Process needs to be started by Chrome)
         /// </summary>
         /// <param name="data">A <see cref="JObject"/> containing the data to be sent.</param>
-        public void SendMessage(JObject data)
-        {
-            Utils.LogMessage("Sending Message:" + JsonConvert.SerializeObject(data));
+        public void SendMessage(JObject data) {
+            Log.LogMessage("Sending Message:" + JsonConvert.SerializeObject(data));
+
             var bytes = System.Text.Encoding.UTF8.GetBytes(data.ToString(Formatting.None));
             var stdout = Console.OpenStandardOutput();
             stdout.WriteByte((byte)((bytes.Length >> 0) & 0xFF));
@@ -95,55 +95,52 @@ namespace NativeMessaging
         /// </summary>
         /// <param name="description">Short application description to be included in the manifest.</param>
         /// <param name="allowedOrigins">List of extensions that should have access to the native messaging host.<br />Wildcards such as <code>chrome-extension://*/*</code> are not allowed.</param>
-        public void GenerateManifest(string description, string[] allowedOrigins)
-        {
-            Utils.LogMessage("Generating Manifest");
-            var manifest = JsonConvert.SerializeObject(new Manifest(Hostname, description, Utils.AssemblyExecuteablePath(), allowedOrigins));
-            File.WriteAllText(ManifestPath, manifest);
-            Utils.LogMessage("Manifest Generated");
-        }
-        /// <summary>
-        /// Register the application to open with Chrome.
-        /// </summary>
-        public void Register()
-        {
-            var regHostnameKeyLocation = RegKeyBaseLocation + Hostname;
+        /// <param name="overwrite">Determines if the manifest should be overwritten if it already exists.<br />Defaults to <see langword="false"/>.</param>
+        public void GenerateManifest(string description, string[] allowedOrigins, bool overwrite = false) {
+            if (File.Exists(ManifestPath) && !overwrite) {
+                Log.LogMessage("Manifest exists already");
+            } else {
+                Log.LogMessage("Generating Manifest");
 
-            RegistryKey regKey = Registry.CurrentUser.OpenSubKey(regHostnameKeyLocation, true);
+                var manifest = JsonConvert.SerializeObject(new Manifest(Hostname, description, Utils.AssemblyExecuteablePath(), allowedOrigins));
+                File.WriteAllText(ManifestPath, manifest);
 
-            if (regKey == null) regKey = Registry.CurrentUser.CreateSubKey(regHostnameKeyLocation);
-
-            regKey.SetValue("", ManifestPath, RegistryValueKind.String);
-
-            regKey.Close();
-            Utils.LogMessage("Registered:" + Hostname);
+                Log.LogMessage("Manifest Generated");
+            }
         }
 
+        #region Browser Registration
         /// <summary>
-        /// Checks if the host is registered with Chrome
+        /// Checks if the host is registered with all required browsers.
         /// </summary>
         /// <returns><see langword="true"/> if the required information is present in the registry.</returns>
-        public bool IsRegisteredWithChrome()
-        {
-            var regHostnameKeyLocation = RegKeyBaseLocation + Hostname;
-            RegistryKey regKey = Registry.CurrentUser.OpenSubKey(regHostnameKeyLocation, true);
+        public bool IsRegistered() {
+            bool result = false;
 
-            if (regKey != null && regKey.GetValue("").ToString() == ManifestPath) return true;
+            foreach (ChromiumBrowser browser in SupportedBrowsers) {
+                result = result || browser.IsRegistered(Hostname, ManifestPath);
+            }
 
-            return false;
+            return result;
         }
 
         /// <summary>
-        /// De-register the application to open with Chrome.
+        /// Register the application to open with all required browsers.
         /// </summary>
-        public void UnRegister()
-        {
-            var regHostnameKeyLocation = RegKeyBaseLocation + Hostname;
-
-            RegistryKey regKey = Registry.CurrentUser.OpenSubKey(regHostnameKeyLocation, true);
-            if (regKey != null) regKey.DeleteSubKey("", true);
-            regKey.Close();
-            Utils.LogMessage("Unregistered:" + Hostname);
+        public void Register() {
+            foreach (ChromiumBrowser browser in SupportedBrowsers) {
+                browser.Register(Hostname, ManifestPath);
+            }
         }
+
+        /// <summary>
+        /// De-register the application to open with all required browsers.
+        /// </summary>
+        public void Unregister() {
+            foreach (ChromiumBrowser browser in SupportedBrowsers) {
+                browser.Unregister(Hostname);
+            }
+        }
+        #endregion
     }
 }
